updated new sql file
Removed unnecessary association relations in favor of foreign keys in other relations
Added tables for multivariate attributes
Added statements to drop public shema, readd it and after tables are created apply all necessary grants. This should simplify updates in this file.
Also removed number of hotels and rooms for relations, these are unnecessary because we can just query the count if need be.
Made hotel_id a unique integer and sole primary key of hotel relation this simplifies referential integrity and has no loss of information.
Made employee effetively an extension of user relation so that we can remove the sign in radio button, this should just handle itself (also simplifies the authenticaiton process.
NOTE: triggers still need to be added for archive.